implemented favourite state & functions
